requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.2
pandas==0.22.0
pkg-resources==0.0.0
pycairo==1.16.3
PyQt5==5.10.1
python-dateutil==2.7.2
pytz==2018.3
pywebview==1.8
requests==2.18.4
sip==4.19.8
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
